wip: viewmodel refactoring
