+++ octoprint_mrbeam/__init__.py
@@ -508,7 +508,7 @@ class MrBeamPlugin(octoprint.plugin.SettingsPlugin,
 			    "js/lasersaftey_viewmodel.js",
 			    "js/ready_to_laser_viewmodel.js",
 			    "js/lib/screenfull.min.js",
-			    "js/calibration/camera_calibration.js",
+			    "js/settings/camera_settings.js",
 			    "js/settings/backlash.js",
 			    "js/settings/leds.js",
 			    "js/path_magic.js",
@@ -533,6 +533,9 @@ class MrBeamPlugin(octoprint.plugin.SettingsPlugin,
 			    "js/settings/custom_material.js",
 			    "js/design_store.js",
 			    "js/settings_menu_navigation.js",
+				"js/calibration/calibration.js"
+				"js/calibration/corner_calibration.js"
+				"js/calibration/lens_calibration.js"
 			    ],
 			css=["css/mrbeam.css",
 			     "css/backlash_settings.css",
