+++ octoprint_mrbeam/templates/calibration/watterott/tab_corner_calibration.jinja2
@@ -37,7 +37,7 @@
             <div class="marker_found_indicator" data-bind="css: { active: camera.markersFound()['SW'] }"><i class="icon-ok"></i> SW</div>
             <div class="marker_found_indicator" data-bind="css: { active: camera.markersFound()['SE'] }"><i class="icon-ok"></i> SE</div>
             <div class="marker_found_indicator" data-bind="css: { active: camera.markersFound()['NE'] }"><i class="icon-ok"></i> NE</div>
-            <div class="marker_found_indicator" data-bind="visible: !cal_img_ready()"><i class="icon-spinner icon-spin"></i> {{ _('detecting') }}</div>
+            <div class="marker_found_indicator" data-bind="visible: !calImgReady()"><i class="icon-spinner icon-spin"></i> {{ _('detecting') }}</div>
         </div>
         <a class="muted" data-bind="attr: { href: cornerCalImgUrl() }"> {{ _('Download this image') }}</a>
     </div>
@@ -46,7 +46,7 @@
 <div data-bind="visible: !cornerCalibrationActive()" width="100%">
     <div style="margin-top:1em;">
         <button class="btn btn-default"
-                data-bind="visible: cal_img_ready() && !cornerCalibrationComplete() && !availablePic()['cropped'], click: startCornerCalibration">
+                data-bind="visible: calImgReady() && !cornerCalibrationComplete() && !availablePic()['cropped'], click: startCornerCalibration">
             {{ _('Start Corner Calibration') }}
         </button>
     </div>
@@ -89,10 +89,10 @@
     <div class="marker_selection_desc calibration_summary">
         <div data-bind="css: {complete: cornerCalibrationComplete()}">
             <div data-bind="visible: cornerCalibrationActive()">
-                <div id="NW" class="cal-row"><div class="cal-arrow NW"></div> North West: <span data-bind="text: __format_point(currentResults()['NW'])"></span></div>
-                <div id="SW" class="cal-row"><div class="cal-arrow SW"></div> South West: <span data-bind="text: __format_point(currentResults()['SW'])"></span></div>
-                <div id="SE" class="cal-row"><div class="cal-arrow SE"></div> South East: <span data-bind="text: __format_point(currentResults()['SE'])"></span></div>
-                <div id="NE" class="cal-row"><div class="cal-arrow NE"></div> North East: <span data-bind="text: __format_point(currentResults()['NE'])"></span></div>
+                <div id="NW" class="cal-row"><div class="cal-arrow NW"></div> North West: <span data-bind="text: _formatPoint(currentResults()['NW'])"></span></div>
+                <div id="SW" class="cal-row"><div class="cal-arrow SW"></div> South West: <span data-bind="text: _formatPoint(currentResults()['SW'])"></span></div>
+                <div id="SE" class="cal-row"><div class="cal-arrow SE"></div> South East: <span data-bind="text: _formatPoint(currentResults()['SE'])"></span></div>
+                <div id="NE" class="cal-row"><div class="cal-arrow NE"></div> North East: <span data-bind="text: _formatPoint(currentResults()['NE'])"></span></div>
             </div>
         </div>
         <div style="margin-top:1em;">
